Improve mashify/rashify specs
The specs used to be isolated from the `call(env)` cycle by means of
using the `on_complete` method directly. Since that method is not
considered part of the public API of Faraday::Response::Middleware, this
switches away from using that method and makes the middleware process
responses the standard way, via `call()`.

This gets rid of the need for "integration tests" because the middleware
is exercised exactly as it would during a real request/response cycle.