Work around Faraday 0.9 query parsing incompatibility in OAuth2
The query string 'access_token&q=hello` gets parsed with the token value
`nil` in Faraday 0.8, but with value `true` in Faraday 0.9.

This is most likely a regression in Faraday, but for now, simply check
if an OAuth token quacks like a string before trying to treat it as one.