Follow redirects that may have an unescaped character in the URL
Even though the URL in the "Location" response header should already be
escaped, somethimes that's not the case. A redirect to a location like
this would cause a Ruby exception:

  /found?action_type_map=["og.likes"]

This is because the double quote characters are not escaped.

Now we escape all characters that cannot be a valid part of an URI, but
we take precaution to not double-escape any existing sequences so we
ignore the "%" character, as it most likely represents something that's
already escaped.

Fixes #87