Fix JSON-related spec failures
Even though "json" was supposed to be in the gem bundle only for Ruby
1.8, Simplecov declares in a json dependency, meaning that `require json`
will never actually manage to load the built-in JSON library on Ruby 1.9.

That's a shame. To work around test inconsistencies, downgrade the
"json" gem to a version compatible with built-in json, so that trying to
parse a JSON body that's a simple value (such as `true` or `1`) properly
raises a parse exception.