Allow `write_options` to be specified for FaradayMiddleware::Caching
`FaradayMiddleware::Caching` is configured with a cache with responds to `#read`, `#fetch` and `#write`. Usually, this will be an instance of a class which conforms to the abstract cache store class [`ActiveSupport::Cache::Store`](http://api.rubyonrails.org/classes/ActiveSupport/Cache/Store.html).

In this pattern, the third argument of `#write` is an optional hash of options - for example, you can specify an `expires_in` to set how long until the cache key expires. Cache stores may support a range of options.

This allows you, when configuring the middleware, to specify a set of `write_options` which will be passed in as this third argument when writing to the cache.

The alternative to this would be passing in some kind of wrapper around your chosen cache store which does this for you, but it feels nicer for the middleware to do it.